pyproject.toml
[tool.poetry]
name = "fastgraph"
version = "0.1.0"
description = "a demo for making a AI search engine using fasthtml and langgraph"
authors = ["Frank Lin <etrobot@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langgraph = "^0.2.14"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.23"
duckduckgo-search = "^6.2.11"
notion-client = "^2.2.1"
python-fasthtml = "^0.5.1"
langgraph-checkpoint-sqlite = "^1.0.3"
shortuuid = "^1.0.13"
aiosqlite = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
